requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antspyx==0.2.9
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
charset-normalizer==2.0.4
chart-studio==1.1.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
dicom2nifti==2.3.0
distlib==0.3.2
docker-pycreds==0.4.0
docstring-parser==0.10
filelock==3.0.12
fsspec==2021.8.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
identify==2.2.13
idna==3.2
imageio==2.9.0
itk==5.2.1
itk-core==5.2.1
itk-filtering==5.2.1
itk-io==5.2.1
itk-numerics==5.2.1
itk-registration==5.2.1
itk-segmentation==5.2.1
joblib==1.0.1
jsonargparse==3.19.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
monai==0.6.0
monai-weekly==0.7.dev2133
multidict==5.1.0
networkx==2.6.2
nibabel==3.2.1
nodeenv==1.6.0
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.2
pathtools==0.1.2
patsy==0.5.1
Pillow==8.3.1
platformdirs==2.2.0
plotly==5.3.1
pre-commit==2.14.0
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pydicom==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.4.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
sentry-sdk==1.3.1
shortuuid==1.0.1
SimpleITK==2.1.0
six==1.16.0
slicer==0.0.7
smmap==4.0.0
statsmodels==0.12.2
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
tifffile==2021.8.30
toml==0.10.2
torch==1.9.0
torchmetrics==0.5.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.7.2
wandb==0.12.1
webcolors==1.11.1
Werkzeug==2.0.1
yarl==1.6.3
